requirements.txt
python-dotenv==1.0.1
transformers==4.40.2
